requirements.txt
torch==1.11.0
torchvision==0.12.0
numpy
tqdm
tensorboard
matplotlib
pandas

speechbrain
datasets
wandb
torchaudio~=0.11.0
pillow
torchmetrics
wget
hparams
gdown
pyworld
unidecode
inflect
librosa